Replace FilterInterceptor with ControllerInterceptor for GWAS plugin
